+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   if it finds any pre-existing secret files. This is helpful when the Secrets
   Provider is being run multiple times.
   [cyberark/secrets-provider-for-k8s#397](https://github.com/cyberark/secrets-provider-for-k8s/pull/397)
+- If the Secrets Provider is run in Push-to-File mode, it no longer errors out
+  if either (a) multiple secret groups use the same secret path, or (b) there
+  are no secrets that need to be retrieved.
+  [cyberark/secrets-provider-for-k8s#404](https://github.com/cyberark/secrets-provider-for-k8s/pull/404)
 
 ## [1.2.0] - 2021-11-30